feature/stylesheet :: A basic css stylesheet, with font and table styling, without the need for bootstrap.
The stylesheet resides at lib/rspec_api_documentation/assets/stylesheets/rspec_api_documentation/styles.css, but may be overridden by creating a stylesheet in the app directory, app/assets/stylesheets/rspec_api_documentation/styles.css.  This allows users to override the default styles if so desired.

The styles are read from the stylesheet and embedded in the HTML, allowing the individual example html pages to be styled without the need for an external resource.